Improved PostgreProjectDAO#findByID method (the SQL query is wrong for the moment)

+++ src/main/java/com/filmplanner/dao/ProjectDAO.java
@@ -7,13 +7,13 @@ public interface ProjectDAO {
 
     Project create(Project project);
 
-    Project findById(int id);
+    Project findById(Long id);
 
     Project[] findManyByManager(User manager);
 
     Project[] findAll();
 
-    void deleteById(int id);
+    void deleteById(Long id);
 
-    void updateById(int id, Project project);
+    void updateById(Long id, Project project);
 }
